pyproject.toml
[tool.poetry]
name = "generate-uml"
version = "0.1.0"
description = "Script to generate airbyte project UML"
authors = ["Aldo Gonzalez <aldo.gonzalez@airbyte.io>"]
[[tool.poetry.packages]]
include = "generate_uml"

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "3.1.43"
graphviz = "0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate-uml = "generate_uml.run:run"